+++ localflavor/ch/forms.py
@@ -57,7 +57,7 @@ class CHIdentityCardNumberField(CharField):
     }
 
     def __init__(self, **kwargs):
-        if "strip" in kwargs and kwargs["strip"] is False:
+        if "strip" in kwargs and kwargs["strip"]:
             raise ImproperlyConfigured("strip cannot be set to False")
         super().__init__(**kwargs)