+++ localflavor/fi/forms.py
@@ -2,9 +2,8 @@
 
 import re
 
-from django.core.validators import EMPTY_VALUES
-from django.forms import ValidationError
-from django.forms.fields import Field, RegexField, Select
+from django.core.exceptions import ImproperlyConfigured, ValidationError
+from django.forms.fields import CharField, RegexField, Select
 from django.utils.translation import gettext_lazy as _
 
 from .fi_municipalities import MUNICIPALITY_CHOICES
@@ -32,17 +31,22 @@ class FIMunicipalitySelect(Select):
         super().__init__(attrs, choices=MUNICIPALITY_CHOICES)
 
 
-class FISocialSecurityNumber(Field):
+class FISocialSecurityNumber(CharField):
     """A form field that validates input as a Finnish social security number."""
 
     default_error_messages = {
         'invalid': _('Enter a valid Finnish social security number.'),
     }
 
+    def __init__(self, **kwargs):
+        if "strip" in kwargs and not kwargs["strip"]:
+            raise ImproperlyConfigured("strip cannot be set to False")
+        super().__init__(**kwargs)
+
     def clean(self, value):
         value = super().clean(value)
-        if value in EMPTY_VALUES:
-            return ''
+        if value in self.empty_values:
+            return value
 
         checkmarks = "0123456789ABCDEFHJKLMNPRSTUVWXY"
         result = re.match(r"""^