+++ localflavor/hr/forms.py
@@ -2,10 +2,9 @@
 import datetime
 import re
 
-from django.core.validators import EMPTY_VALUES
+from django.core.exceptions import ImproperlyConfigured
 from django.forms import ValidationError
-from django.forms.fields import Field, RegexField, Select
-from django.utils.encoding import force_str
+from django.forms.fields import CharField, RegexField, Select
 from django.utils.translation import gettext_lazy as _
 from stdnum import luhn
 
@@ -35,7 +34,7 @@ class HRLicensePlatePrefixSelect(Select):
         super().__init__(attrs, choices=HR_LICENSE_PLATE_PREFIX_CHOICES)
 
 
-class HRJMBGField(Field):
+class HRJMBGField(CharField):
     """
     Unique Master Citizen Number (JMBG) field.
 
@@ -56,12 +55,15 @@ class HRJMBGField(Field):
         'date': _('Error in date segment'),
     }
 
+    def __init__(self, **kwargs):
+        if "strip" in kwargs and not kwargs["strip"]:
+            raise ImproperlyConfigured("strip cannot be set to False")
+        super().__init__(**kwargs)
+
     def clean(self, value):
         value = super().clean(value)
-        if value in EMPTY_VALUES:
-            return ''
-
-        value = value.strip()
+        if value in self.empty_values:
+            return value
 
         matches = jmbg_re.search(value)
         if matches is None:
@@ -117,7 +119,7 @@ class HROIBField(RegexField):
         return '%s' % (value, )
 
 
-class HRLicensePlateField(Field):
+class HRLicensePlateField(CharField):
     """
     Vehicle license plate of Croatia field.
 
@@ -139,12 +141,17 @@ class HRLicensePlateField(Field):
         'number': _('Number part cannot be zero'),
     }
 
+    def __init__(self, **kwargs):
+        if "strip" in kwargs and not kwargs["strip"]:
+            raise ImproperlyConfigured("strip cannot be set to False")
+        super().__init__(**kwargs)
+
     def clean(self, value):
         value = super().clean(value)
-        if value in EMPTY_VALUES:
-            return ''
+        if value in self.empty_values:
+            return value
 
-        value = re.sub(r'[\s\-]+', '', force_str(value.strip())).upper()
+        value = re.sub(r'[\s\-]+', '', value).upper()
 
         matches = plate_re.search(value)
         if matches is None:
@@ -163,7 +170,7 @@ class HRLicensePlateField(Field):
         return '%s %s-%s' % (prefix, number, matches.group('suffix'))
 
 
-class HRPostalCodeField(Field):
+class HRPostalCodeField(CharField):
     """
     Postal code of Croatia field.
 
@@ -176,12 +183,16 @@ class HRPostalCodeField(Field):
         'invalid': _('Enter a valid 5 digit postal code'),
     }
 
+    def __init__(self, **kwargs):
+        if "strip" in kwargs and not kwargs["strip"]:
+            raise ImproperlyConfigured("strip cannot be set to False")
+        super().__init__(**kwargs)
+
     def clean(self, value):
         value = super().clean(value)
-        if value in EMPTY_VALUES:
-            return ''
+        if value in self.empty_values:
+            return value
 
-        value = value.strip()
         if not postal_code_re.search(value):
             raise ValidationError(self.error_messages['invalid'], code='invalid')
 
@@ -192,7 +203,7 @@ class HRPostalCodeField(Field):
         return '%s' % value
 
 
-class HRJMBAGField(Field):
+class HRJMBAGField(CharField):
     """
     Unique Master Academic Citizen Number of Croatia (JMBAG) field.
 
@@ -206,12 +217,17 @@ class HRJMBAGField(Field):
         'copy': _('Card issue number cannot be zero'),
     }
 
+    def __init__(self, **kwargs):
+        if "strip" in kwargs and not kwargs["strip"]:
+            raise ImproperlyConfigured("strip cannot be set to False")
+        super().__init__(**kwargs)
+
     def clean(self, value):
         value = super().clean(value)
-        if value in EMPTY_VALUES:
-            return ''
+        if value in self.empty_values:
+            return value
 
-        value = re.sub(r'[\-\s]', '', value.strip())
+        value = re.sub(r'[\-\s]', '', value)
 
         matches = jmbag_re.search(value)
         if matches is None: