+++ localflavor/il/forms.py
@@ -1,9 +1,8 @@
 """Israeli-specific form helpers."""
 import re
 
-from django.core.exceptions import ValidationError
-from django.core.validators import EMPTY_VALUES
-from django.forms.fields import Field, RegexField
+from django.core.exceptions import ImproperlyConfigured, ValidationError
+from django.forms.fields import CharField, RegexField
 from django.utils.translation import gettext_lazy as _
 from stdnum import luhn
 
@@ -30,7 +29,7 @@ class ILPostalCodeField(RegexField):
         return super().clean(value)
 
 
-class ILIDNumberField(Field):
+class ILIDNumberField(CharField):
     """
     A form field that validates its input as an Israeli identification number.
 
@@ -52,11 +51,16 @@ class ILIDNumberField(Field):
         'invalid': _('Enter a valid ID number.'),
     }
 
+    def __init__(self, **kwargs):
+        if "strip" in kwargs and not kwargs["strip"]:
+            raise ImproperlyConfigured("strip cannot be set to False")
+        super().__init__(**kwargs)
+
     def clean(self, value):
         value = super().clean(value)
 
-        if value in EMPTY_VALUES:
-            return ''
+        if value in self.empty_values:
+            return value
 
         match = id_number_re.match(value)
         if not match: