+++ localflavor/ro/forms.py
@@ -1,8 +1,8 @@
 """Romanian specific form helpers."""
 import datetime
 
-from django.core.validators import EMPTY_VALUES
-from django.forms import Field, RegexField, Select, ValidationError
+from django.core.exceptions import ImproperlyConfigured
+from django.forms import CharField, RegexField, Select, ValidationError
 from django.utils.translation import gettext_lazy as _
 
 from .ro_counties import COUNTIES_CHOICES
@@ -114,7 +114,7 @@ class ROCNPField(RegexField):
         return value
 
 
-class ROCountyField(Field):
+class ROCountyField(CharField):
     """
     A form field that validates its input is a Romanian county name or abbreviation.
 
@@ -137,16 +137,18 @@ class ROCountyField(Field):
         'invalid': 'Enter a Romanian county code or name.',
     }
 
+    def __init__(self, **kwargs):
+        if "strip" in kwargs and not kwargs["strip"]:
+            raise ImproperlyConfigured("strip cannot be set to False")
+        super().__init__(**kwargs)
+
     def clean(self, value):
         value = super().clean(value)
 
-        if value in EMPTY_VALUES:
-            return ''
+        if value in self.empty_values:
+            return value
 
-        try:
-            value = value.strip().upper()
-        except AttributeError:
-            pass
+        value = value.upper()
 
         # search for county code
         for entry in COUNTIES_CHOICES: